+++ build.rs
@@ -2,6 +2,6 @@ extern crate cc;
 
 fn main() {
     cc::Build::new()
-        .file("src/signal.c")
-        .compile("signal");
+        .file("src/impl.c")
+        .compile("impl");
 }