+++ CHANGELOG.md
@@ -4,7 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## 0.1.1 - 2018-03-28
+## [Unreleased]
+
+## [0.2.0] - 2018-03-29
+### Added
+- Add store functions.
+
+### Changed
+- `bulletproof` is now thread-safe (#2).
+- Do not uninstall the installed signal handler (#3).
+
+## [0.1.1] - 2018-03-28
 ### Changed
 - Fix Cargo.toml categories.
 
@@ -12,5 +22,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - First version of bulletproof load.
 
-[Unreleased]: https://github.com/jeehoonkang/bulletproof-rs/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/jeehoonkang/bulletproof-rs/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/jeehoonkang/bulletproof-rs/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/jeehoonkang/bulletproof-rs/compare/v0.1.0...v0.1.1